+++ src/utils.ts
@@ -67,6 +67,8 @@ function hasAnyOf(object: any, props: PayloadKey[]): boolean {
  */
 
 function isEqual<T>(a: any, b: T, depth = 1): a is T {
+  if (a instanceof Date && b instanceof Date) return a.getTime() === b.getTime()
+
   if (!a || !b || (typeof a !== 'object' && typeof b !== 'object'))
     return a === b